Activate, Deactivate pages plus work on paging course list

+++ src/ISIS.ReadModel/IReadRepository.cs
@@ -8,13 +8,13 @@ namespace ISIS
     public interface IReadRepository : IDisposable
     {
 
-        PageOf<T> All<T>(int pageSize, long pageNumber) 
+        PageOf<T> All<T>(int pageSize, int pageNumber) 
             where T : class, IEntity;
 
-        PageOf<T> All<T>(int pageSize, long pageNumber, params Expression<Func<T, object>>[] orderBys)
+        PageOf<T> All<T>(int pageSize, int pageNumber, params Expression<Func<T, object>>[] orderBys)
             where T : class, IEntity;
 
-        PageOf<T> All<T>(int pageSize, long pageNumber, bool descending, Expression<Func<T, object>> orderBy)
+        PageOf<T> All<T>(int pageSize, int pageNumber, bool descending, Expression<Func<T, object>> orderBy)
             where T : class, IEntity;
 
         T Single<T>(object id) where T : IEntity;
